--- lib/src/utils/color_swatches.dart
@@ -1,103 +1,0 @@
-import 'package:flutter/material.dart' show MaterialColor;
-import 'colors.dart';
-import 'hex.dart';
-
-class ColorsSwatches {
-  const ColorsSwatches._();
-
-  static const orange = MaterialColor(
-    YaruHex.ubuntuOrange,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.ubuntuOrange,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-
-  static const lightAubergine = MaterialColor(
-    YaruHex.lightAubergine,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.lightAubergine,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-
-  static const midAubergine = MaterialColor(
-    YaruHex.midAubergine,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.midAubergine,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-
-  static const canonicalAubergine = MaterialColor(
-    YaruHex.canonicalAubergine,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.canonicalAubergine,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-
-  static const darkAubergine = MaterialColor(
-    YaruHex.darkAubergine,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.darkAubergine,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-
-  static const warmGrey = MaterialColor(
-    YaruHex.darkAubergine,
-    {
-      50: YaruColors.roseWhite,
-      100: YaruColors.pippin,
-      200: YaruColors.melon,
-      300: YaruColors.darkSalmon,
-      400: YaruColors.burntSienna,
-      500: YaruColors.warmGrey,
-      600: YaruColors.rust,
-      700: YaruColors.burntUmber,
-      800: YaruColors.bakersChocolate,
-      900: YaruColors.sealBrown,
-    },
-  );
-}
