--- lib/src/utils/colors.dart
@@ -1,32 +1,0 @@
-import 'dart:ui';
-
-import 'hex.dart';
-
-// Ubuntu color palette `https://design.ubuntu.com/brand/colour-palette/`
-
-class YaruColors {
-  const YaruColors._();
-  static const bakersChocolate = Color(YaruHex.bakersChocolate);
-  static const black = Color(YaruHex.black);
-  static const burntSienna = Color(YaruHex.burntSienna);
-  static const burntUmber = Color(YaruHex.burntUmber);
-  static const canonicalAubergine = Color(YaruHex.canonicalAubergine);
-  static const coolGrey = Color(YaruHex.coolGrey);
-  static const darkAubergine = Color(YaruHex.darkAubergine);
-  static const darkSalmon = Color(YaruHex.darkSalmon);
-  static const disabledGreyDark = Color(YaruHex.disabledGreyDark);
-  static const errorColor = Color(YaruHex.errorColor);
-  static const green = Color(YaruHex.green);
-  static const lightAubergine = Color(YaruHex.lightAubergine);
-  static const melon = Color(YaruHex.melon);
-  static const midAubergine = Color(YaruHex.midAubergine);
-  static const pippin = Color(YaruHex.pippin);
-  static const red = Color(YaruHex.red);
-  static const roseWhite = Color(YaruHex.roseWhite);
-  static const rust = Color(YaruHex.rust);
-  static const sealBrown = Color(YaruHex.sealBrown);
-  static const textGrey = Color(YaruHex.textGrey);
-  static const ubuntuOrange = Color(YaruHex.ubuntuOrange);
-  static const warmGrey = Color(YaruHex.warmGrey);
-  static const white = Color(YaruHex.white);
-}
