+++ lib/src/utils/swatch.dart
@@ -1,0 +1,103 @@
+import 'package:flutter/material.dart' show MaterialColor;
+import 'color.dart';
+import 'hex.dart';
+
+class YaruSwatch {
+  const YaruSwatch._();
+
+  static const orange = MaterialColor(
+    YaruHex.ubuntuOrange,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.ubuntuOrange,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+
+  static const lightAubergine = MaterialColor(
+    YaruHex.lightAubergine,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.lightAubergine,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+
+  static const midAubergine = MaterialColor(
+    YaruHex.midAubergine,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.midAubergine,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+
+  static const canonicalAubergine = MaterialColor(
+    YaruHex.canonicalAubergine,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.canonicalAubergine,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+
+  static const darkAubergine = MaterialColor(
+    YaruHex.darkAubergine,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.darkAubergine,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+
+  static const warmGrey = MaterialColor(
+    YaruHex.darkAubergine,
+    {
+      50: YaruColor.roseWhite,
+      100: YaruColor.pippin,
+      200: YaruColor.melon,
+      300: YaruColor.darkSalmon,
+      400: YaruColor.burntSienna,
+      500: YaruColor.warmGrey,
+      600: YaruColor.rust,
+      700: YaruColor.burntUmber,
+      800: YaruColor.bakersChocolate,
+      900: YaruColor.sealBrown,
+    },
+  );
+}
