+++ lib/yaru.dart
@@ -1,4 +1,6 @@
 library yaru;
 
+export 'package:yaru/src/utils/color.dart';
 export 'package:yaru/src/theme.dart';
-export 'package:yaru/src/utils/colors.dart';
+export 'package:yaru/src/utils/hex.dart';
+export 'package:yaru/src/utils/swatch.dart';
